add extras/todone-view_sh - a shell implementation of todone-view
This script relies on non-POSIX arguments to `date` as well as the presence of
either `tac` (Linux) or `tail -r` (BSD). It's also slower than the Lua
implementation, but given the output is piped to a pager by default and users
aren't likely to look far into the past the speed difference is not important.