fix bug arising from last commit
I didn't rm todo.tmp after todone fails to find a match which means it was
appended to the next time todone was run successfully. This really
demonstrates another bug - the use of touch in todone and todone-archive. I
want the file to exist, but I also want it to be zero-length. Therefore don't
use touch and just use file redirection to create or truncate the file.